Added vehicle status images to leaderboard rows

+++ src/apps/rNascar23-Maui/rNascar23Multi/Pages/LeaderboardPage.xaml.cs
@@ -149,7 +149,7 @@ public partial class LeaderboardPage : ContentPage, INotifyPropertyChanged
             // TODO Speed versus time
             model.BestLap = orderedVehicles[i].BestLapSpeed;
             model.OnLap = orderedVehicles[i].BestLap;
-            model.Status = (int)orderedVehicles[i].Status;
+            model.Status = orderedVehicles[i].Status;
         }
     }
